+++ backend/controllers/address.py
@@ -1,12 +1,10 @@
-import json
-
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
 from rest_framework.parsers import JSONParser
 from backend.models import Address, Tag
 from backend.serializers import AddressSerializer, TagSerializer
 
-from backend.controllers.view_utils import get_authenticated_user
+from backend.controllers.controller_utils import *
 
 
 @csrf_exempt
@@ -20,9 +18,9 @@ def address_detail(request, pk):
     try:
         address = Address.objects.get(pk=pk)
     except Address.DoesNotExist:
-        return HttpResponse(status=404)
+        return HttpResponse(status=HTTP_STATUS_NOT_FOUND)
 
-    response = HttpResponse(status=405)
+    response = HttpResponse(status=HTTP_STATUS_METHOD_NOT_ALLOWED)
     if request.method == 'GET':
         response = __retrieve_address(address)
     elif request.method == 'DELETE':
@@ -38,7 +36,7 @@ def address_list(request):
     :param request:
     :return: list of addresses
     """
-    response = HttpResponse(status=405)
+    response = HttpResponse(status=HTTP_STATUS_METHOD_NOT_ALLOWED)
     if request.method == "GET":
         response = __list_addresses()
     elif request.method == "POST":
@@ -57,9 +55,9 @@ def address_tags(request, pk):
     try:
         address = Address.objects.get(pk=pk)
     except Address.DoesNotExist:
-        return HttpResponse(status=404)
+        return HttpResponse(status=HTTP_STATUS_NOT_FOUND)
 
-    response = HttpResponse(status=405)
+    response = HttpResponse(status=HTTP_STATUS_METHOD_NOT_ALLOWED)
     if request.method == 'GET':
         response = __retrieve_tags_by_address(address)
 
@@ -74,11 +72,11 @@ def __retrieve_address(address):
 def __delete_address(request, address):
     authentication_request = get_authenticated_user(request)
 
-    if not authentication_request.status_code == 200:
+    if not authentication_request.status_code == HTTP_STATUS_OK:
         return authentication_request
 
     address.delete()
-    return HttpResponse(status=204)
+    return HttpResponse(status=HTTP_STATUS_NO_CONTENT)
 
 
 def __list_addresses():
@@ -90,23 +88,25 @@ def __list_addresses():
 def __create_address(request):
     authentication_request = get_authenticated_user(request)
 
-    if not authentication_request.status_code == 200:
+    if not authentication_request.status_code == HTTP_STATUS_OK:
         return authentication_request
 
     authenticated_data = json.loads(authentication_request.content.decode('utf-8'))
     request_data = json.loads(request.body.decode('utf-8'))
 
     if not authenticated_data['id'] == request_data['creator_user_id']:
-        return JsonResponse({"Error": authenticated_data["id"] + " " + request_data['creator_user_id']}, status=403)
+        return JsonResponse({"Error": authenticated_data["id"] + " " + request_data['creator_user_id']},
+                            status=HTTP_STATUS_FORBIDDEN)
 
     data = JSONParser().parse(request)
     serializer = AddressSerializer(data=data)
     if serializer.is_valid():
         serializer.save()
-        return JsonResponse(serializer.data, status=201)
-    return JsonResponse(serializer.errors, status=400)
+        return JsonResponse(serializer.data, status=HTTP_STATUS_CREATED)
+    return JsonResponse(serializer.errors, status=HTTP_STATUS_BAD_REQUEST)
 
 
 def __retrieve_tags_by_address(address):
-    tags = Tag.objects.filter(address=address    serializer = TagSerializer(tags, many=True)
+    tags = Tag.objects.filter(address=address)
+    serializer = TagSerializer(tags, many=True)
     return JsonResponse(serializer.data, safe=False)
