+++ backend/controllers/controller_utils.py
@@ -4,6 +4,15 @@ import json
 from django.http import JsonResponse
 
 
+HTTP_STATUS_OK = 200
+HTTP_STATUS_CREATED = 201
+HTTP_STATUS_NO_CONTENT = 204
+HTTP_STATUS_BAD_REQUEST = 400
+HTTP_STATUS_FORBIDDEN = 403
+HTTP_STATUS_NOT_FOUND = 404
+HTTP_STATUS_METHOD_NOT_ALLOWED = 405
+
+
 def get_authenticated_user(request):
     auth = request.META.get("HTTP_AUTHORIZATION", None)
