+++ backend/controllers/csv.py
@@ -1,7 +1,9 @@
 from django.http import HttpResponse
 from djqscsv import render_to_csv_response
 from django.views.decorators.csrf import csrf_exempt
+
 from backend.models import Address, Tag
+from backend.controllers.controller_utils import HTTP_STATUS_METHOD_NOT_ALLOWED
 
 
 @csrf_exempt
@@ -11,7 +13,7 @@ def csv_address_export(request):
     :param request:
     :return: a .csv file for download
     """
-    response = HttpResponse(status=405)
+    response = HttpResponse(status=HTTP_STATUS_METHOD_NOT_ALLOWED)
     if request.method == "GET":
         addresses = Address.objects.all()
         response = render_to_csv_response(queryset=addresses, filename='addresses.csv', append_datestamp=True)
@@ -25,7 +27,7 @@ def csv_tag_export(request):
     :param request:
     :return: a .csv file for download
     """
-    response = HttpResponse(status=405)
+    response = HttpResponse(status=HTTP_STATUS_METHOD_NOT_ALLOWED)
     if request.method == "GET":
         tags = Tag.objects.all()
         response = render_to_csv_response(queryset=tags, filename='tags.csv', append_datestamp=True)
